chore: simplify object ordering
